pyproject.toml
[tool.poetry]
name = "raster2dggs"
version = "0.6.0"
description = ""
authors = ["James Ardo <ardoj@landcareresearch.co.nz>"]
maintainers = ["Richard Law <lawr@landcareresearch.co.nz>"]
readme = "README.md"
license = "LGPL-3.0-or-later"
repository = "https://github.com/manaakiwhenua/raster2dggs"
keywords = ["dggs", "raster", "h3", "rHEALPix", "cli"]
classifiers = [
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Information Analysis"
]

[tool.poetry.dependencies]
python = "^3.11"
gdal = "^3.8"
geopandas = "^1.0"
h3pandas = "^0.3"
rioxarray = "^0.19"
dask-geopandas = "^0.5"
pyarrow = "^20.0"
dask = "^2025.1"
click = "^8.1"
boto3 = "^1"
tqdm = "^4"
click-log = "^0.4"
rasterio = "^1.4"
dask-expr = "^2"
numpy = "^2"
rhppandas = "^0.1"
rhealpixdggs = "^0.5"
python-geohash = "^0.8"
maidenhead = "^1.8"
s2sphere = "^0.2"
pya5 = "^0.5"
geoarrow-pyarrow = "^0.2.0"
dggal = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
twine ="*"
black = "*"

[tool.poetry.scripts]
raster2dggs = "raster2dggs.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88